backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.4

# AI / LangGraph (pinned to compatible set)
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.7

# Google GenAI SDK (new, non-deprecated)
google-genai>=0.7.0

# Database
supabase==2.4.6

# Telegram
python-telegram-bot==21.3

# Crop Journey
feedparser>=6.0.0

# Optional: NeMo Guardrails (install separately if needed)
# nemoguardrails==0.9.0
